pyproject.toml
[tool.poetry]
name = "world-dash"
version = "0.1.0"
description = "Geopolitical Intelligence Dashboard - Real-time world events monitoring"
authors = ["World Dash Team"]
readme = "README.md"
packages = [
    { include = "packages/shared" },
    { include = "packages/feed_ingestion" },
    { include = "packages/event_normalizer" },
    { include = "packages/intelligence_engine" },
    { include = "packages/storage" },
    { include = "apps/api" },
    { include = "apps/worker" },
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
geoalchemy2 = "^0.14.3"
redis = "^5.0.1"
celery = "^5.3.6"
feedparser = "^6.0.11"
httpx = "^0.26.0"
geopy = "^2.4.1"
python-dateutil = "^2.8.2"
structlog = "^24.1.0"
prometheus-client = "^0.19.0"
python-multipart = "^0.0.6"
newspaper3k = "^0.2.8"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
ruff = "^0.1.14"
black = "^24.1.1"
mypy = "^1.8.0"
httpx = "^0.26.0"
faker = "^22.6.0"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
